Fixing navbar and tutorials
Made navbar be on every page and added sign in/profile button depending on whether or not the user is logged in. The tutorials page is similar to the projects page but each tutorial card is just a link to the actual tutorial. Also the tutorials are fetched from dynamodb using aws lambda.